// WebApp-Compiler-FRONTEND/src/app/models/prg-question-model.ts
export class PrgQuestionModel 
{
  id: number;
  questionId: string;
  solution: string;
  basic: string;
  language: string;
  inputs: string;
  testcase1: string;
  testcase1A: string;
  testcase2: string;
  testcase2A: string;
  testcase3: string;
  testcase3A: string;
  testcase4: string;
  testcase4A: string;
  testcase5: string;
  testcase5A: string;
  testcase6: string;
  testcase6A: string;
  testcase7: string;
  testcase7A: string;
  testcase8: string;
  testcase8A: string;
  testcase9: string;
  testcase9A: string;
  testcase10: string;
  testcase10A: string;
}